refactor: rename grid state file
